fixed benchmark test crash - xfield history should be initialized in the test beginning. changed initial value of block sizelimit to 1MB instead of 1M-1000
