+++ src/bench/bench_tapyrus.cpp
@@ -4,11 +4,13 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <bench/bench.h>
+#include <bench/bench_genesis.h>
 
 #include <crypto/sha256.h>
 #include <key.h>
 #include <utilstrencodings.h>
 #include <validation.h>
+#include <streams.h>
 
 #include <memory>
 
@@ -44,21 +46,7 @@ static fs::path SetDataDir()
     gArgs.ForceSetArg("-datadir", ret.string());
     return ret;
 }
-//67ae3f5bfb3464b9704d7bd3a134401cc80c3a172240ebfca9f1e40f51bb6d37
-void writeTestGenesisBlockToFile(fs::path genesisPath)
-{
-    genesisPath /= TAPYRUS_GENESIS_FILENAME;
-    //printf("Writing Genesis Block to [%s]\n", genesisPath.string().c_str());
-    fs::ofstream stream(genesisPath);
-    CKey aggregatePrivKey;
-    const unsigned char validAggPrivateKey[32] = {0x67, 0xae, 0x3f, 0x5b, 0xfb, 0x34, 0x64, 0xb9, 0x70, 0x4d, 0x7b, 0xd3, 0xa1, 0x34, 0x40, 0x1c, 0xc8, 0x0c, 0x3a, 0x17, 0x22, 0x40, 0xeb, 0xfc, 0xa9, 0xf1, 0xe4, 0x0f, 0x51, 0xbb, 0x6d, 0x37};
-    aggregatePrivKey.Set(validAggPrivateKey, validAggPrivateKey + 32, true);
-    CBlock genesis(createGenesisBlock(aggregatePrivKey.GetPubKey(), aggregatePrivKey));
-    CDataStream ss(SER_NETWORK, PROTOCOL_VERSION);
-    ss << genesis;
-    stream << HexStr(ss.begin(), ss.end());
-    stream.close();
-}
+
 int main(int argc, char** argv)
 {
     SetupBenchArgs();
@@ -80,7 +68,7 @@ int main(int argc, char** argv)
     SHA256AutoDetect();
     ECC_Start();
     SetupEnvironment();
-    writeTestGenesisBlockToFile(bench_datadir);
+    writeTestGenesisBlockToFile(GetDataDir(false));
     SelectFederationParams(TAPYRUS_OP_MODE::PROD);
     CXFieldHistory xFieldHistory(FederationParams().GenesisBlock());
